teste backup

+++ Miguel/v3.0_TCP_UDP/missoes.py
@@ -23,7 +23,7 @@ def int_to_mission(missao: int) -> str:
     raise ValueError(f"ID de missão desconhecido: {missao}")
 
 
-def updateWork(valor: int, proAtual: float, dur: int) -> float:
+def updateWork(valor: int, proAtual: int, dur: int) -> int:
     incremento = {
         0: 0,
         1: 2,
@@ -36,12 +36,10 @@ def updateWork(valor: int, proAtual: float, dur: int) -> float:
 
     if valor not in incremento:
         raise ValueError("Valor fora do intervalo permitido (0-6)")
-
+    
     inc_base = incremento[valor]
-    inc_escalado = inc_base * (dur / 60.0)
-
-    # acumula SEM truncar
-    return min(proAtual + inc_escalado, 100.0)
+    inc_escalado = inc_base * (dur / 60)
+    return min(int(proAtual + inc_escalado), 100)
